Intent: decouple window validation for coherent reuse Operators involved: [Coherence, Recursivity] Affected invariants: [#4, #8] Key changes: - Extract validate_window into tnfr.validation.window for dependency-light access - Update tnfr.validation exports and glyph_history to use the new module - Refresh typing stubs to reflect the refactor Expected risks/dissonances: Import churn in validation package; mitigated with re-exports and tests Metrics: C(t) unchanged; νf unaffected; phase stable; ΔNFR neutral Equivalence map: tnfr.validation.validate_window now delegates to tnfr.validation.window.validate_window
