+++ src/tnfr/validation/graph.pyi
@@ -14,6 +14,4 @@ AliasSequence = Sequence[str]
 
 GRAPH_VALIDATORS: Tuple[ValidatorFunc, ...]
 
-def validate_window(window: int, *, positive: bool = ...) -> int: ...
-
 def run_validators(graph: TNFRGraph) -> None: ...
