+++ src/tnfr/validation/window.pyi
@@ -1,0 +1,1 @@
+def validate_window(window: int, *, positive: bool = ...) -> int: ...
